fastapi/requirements.txt
alembic==1.14.0
annotated-types==0.6.0
anyio==3.7.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.8
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.105.0
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.6
joblib==1.4.0
Levenshtein==0.25.0
Mako==1.3.8
MarkupSafe==3.0.2
numpy==1.26.4
openpyxl==3.1.2
osm2geojson==0.2.5
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
psycopg2==2.9.9
py-bcrypt==0.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pypdf==4.2.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-Levenshtein==0.25.0
python-multipart==0.0.6
pytz==2024.1
rapidfuzz==3.8.0
reportlab==4.1.0
requests==2.31.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
threadpoolctl==3.4.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.24.0.post1
virtualenv==20.25.0
